Complete initial rendering of the space

+++ src/modules/consts.rs
@@ -1,3 +1,4 @@
+#![allow(dead_code)]
 use std::f32::consts::PI;
 
 pub const PI_2: f32 = PI / 2.;
@@ -10,8 +11,8 @@ pub const EYE_Z: f32 = 1.65;
 pub const HFOV: f32 = 90.0 * PI / 180.0;
 pub const VFOV: f32 = 0.5;
 
-pub const MAP_SIZE: usize = 8;
-pub const MAP_DATA: [[u32; MAP_SIZE]; MAP_SIZE] = [
+pub const MAP_SIZE: i32 = 8;
+pub const MAP_DATA: [[i32; MAP_SIZE as usize]; MAP_SIZE as usize] = [
     [1, 1, 1, 1, 1, 1, 1, 1],
     [1, 0, 0, 0, 0, 0, 0, 1],
     [1, 0, 0, 0, 0, 3, 0, 1],
